requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
prophet>=1.1.0
tensorflow>=2.8.0
flask>=2.0.0
pyyaml>=6.0
requests>=2.27.0
prometheus-client>=0.13.0
influxdb-client>=1.28.0
python-dateutil>=2.8.0
